smurf: skyloop: New parameter INITIALSKY allows an initial sky map to be given
The map supplied via parameter INITIALSKY defines the starting point for
the iterative map-maker algorithm, instead of the usual flat map full of
zeros. It also defines the output map's WCS and pixel bounds.

The motivation for this was to see if using the 850 um Q/U maps as the
initial sky when creating the 450um Q/U maps may remove the angular
offset sometimes seen between 450 and 850 vectyor maps (e.g. in the
BISTRO Oph A field).