kappa: sqorst: new parameter CENTRE specifies centre of squash or stretch
Previously sqorst always squashed or stretched about the centre of the
input array. This resulted in the pixel origin changing its position on
the sky (unless the origin was also the array centre). The new parameter
allows the squashing or stretching to be performed about the pixel
origin, so that the pixel origin retains its position on the sky. The
default is to squash or stretch about the array centre.