smurf: makemap: Calculate and report the ch-squared in the source regions
This is the weighted mean of the squared time-stream residuals divided by
the variance of the corresponding time-stream data. It gives a measure of
the closeness of fit of the time-stream data implied by the final map
(and other models) to the supplied time-stream data. The weight for each
time-stream value in this mean is the SNR of the final map pixel in which
the time-stream value falls. This means that the value primarily
represents the chi-squared in the source regions (i.e. the regions of
high SNR), and is not heavily influenced by the chi-squared in the
background regions, which is usually much closer to 1.0 than in the
source regions.