ndf: Allow WCS section extents to be specified by arc-distance
It has been possible to specify NDF section using WCS coordinates for
several years, but there has always been a problem specifying an RA
section using the "centre~extent" format because an extent in RA does not
correspond to an arc-distance (unlike an extent in Dec), but varies in
size depending on the Declination.

This commit allows the extent to be flagged as an arc-distance
explicitly. For instance image(10:12:30~40am,-23:23:43~40am)' will create
a section of extent 40 arc-minutes on both axes ("as" and "ad" can be used
in place of "am", indicating arc-seconds and arc-degrees).

This only works if the NDFs current Frame is a SkyFrame (an error is
reported otherwise).