smurf: Add a third opacity relation coefficient
This allows for a third (sqrt) term in the SCUBA-2 filter opacity
relations.  The smf_cso2filt_coeff function now provides default
values for unspecified coefficients so that existing configurations
(which do not specify a third value) use a coefficient of zero for
the new term.  (The default values are taken from a literal array
in case future coefficients require a non-zero default, such as
the power of the new term.)