smurf: skyloop: Reduce chances of error in array import
On the second and subsequent iterations, makemap imports models (such as
EXT) created on the first iteration. If such a model contains many bad
values, there is an option to fill those bad values with the mean of the
nearby time slices. Here, "nearby" means the time slices processed by the
same worker thread. If a machine has many cores, the section of time
slices processed by a single core will be small, thus increasing the
chances that no good time slice will be found within the section, causing
makemap and skyloop to abort. This commit limits the number of worker
threads to three, so each thread will get a third of the entire time
stream, increasing the chances of at least one good time slice being
found.