HDSTRACE: Fixed bug that caused values not to appear in wide terminals.
The maximum number of lines of values was set such that it generated
a maximum number of characters for listing the values at 128.  This
arose because of the too-low 64 in its calculation.  When computing
the number of values for HDS to extract for wider terminals (c.150
characters, depending on the object and indentation), this scaling
lead to a wrap around of very large integers, and consequently asking
HDS to return a negative number of data values.  In turn, the bad
status led to '{undefined}' appearing in the output.

This factor is now increased to the maximum width allowed by the
message system, less ten for a minimum object name and dimensions, and
indentation.  The effective screen width is capped to MSG__SZMSG.
This maximum output width is now noted in the documentation.

The original choice of 64 was due to HDSTRACE's genesis being at a
time when printers and terminals had a maximum width of 132
characters.