smurf: skyloop: fix bug in use of masks created on 1st iteration
The first use of makemap creates an output map in which the bad-bits mask
is set to 255 (i.e. all masks switched on). This map is used later to
calculate the normalised map change introduced by the second use of
makemap. But that map-change estimate should be based on the AST mask
generated by the second makemap. With things as they were, it would also
be affected by the masks created by the first use of makemap. Forcing the
bad-bits value in the initial map to zero (i.e. all masks switched off)
means that the pixels used to calculate the map-change estimate will be
selected soley on the basis of the AST mask in the map created by the
second use of makemap.