smurf: makemap: New config param to omit initial time slices
Looks like the initial section of time slices, whilst the telescope is
accelerating and getting into the requirerd scan pattern, may produce
artifical extended features in the final map. So allow a specified number
of time slices to be ignored at the start of an observation. See new
config parameter "startup".