smurf: calcqu: Fix bug that caused bad output RTS_NUM and STEPTIME values
The RTS_NUM values in the JCMTSTATE extension of the output time streams
created by calcqu were inaccurate if the sample rate in the input data is
not constant. This bug was not too serious as the RTS_NUM values are not
used for anything significant in makemap. The same bug caused the
STEPTIME FITS header to be inaccurate.