smurf: configmeld: Allow use with commands other than makemap
configmeld now has a new parameter called APP that can be set to the name
of the command for which config parameters should be displayed. For instance
to see the config supplied by the user to the skyloop command, run with
"app=skyloop" (otherwise the config used by makemap will be displayed).