smurf: makemap: Allow COM to be dumped a a full 2D map
If diag.map was set, the COM model was previously dumped as a 2D
representation of a single bolometer time stream representing the common
mode at each time slice. Now it is dumped as the full 2D map including
all bolometers. The COM time-stream is scaled using the GAI model to
create the comm-mode for each bolometer.