SC/20:  Give examples of both RECPTORS and RECPUSED keywords.
There was a paragraph and example that was repeated verbatim in
Sections 6.4 and 7.2.3.

Rewrote the earlier one to use the RECPTORS keyword, but noted the
role of RECPUSED, with a link to the later paragraph.

In the later paragraph, note that RECPUSED records the good receptors
used by MAKECUBE to assemble the spectral cube.  Add a HARP example
of a group reduced cube.

Also inserted a hyperlink link to FITSVAL, wrapped \param{} around
keywords, and used em-dashes.