SC/21: List all the source LaTeX files as dependencies.
This is so that if any one text component of the document is revised,
a "make" will rebuild the whole manual rather than doing nothing.

Removed the sc21_htx_tar dependencies, not least because the list of
graphics files was out of date, and in the rare occurence of a
graphics file changing but retaining its name, it's easy enough to
remove sc21.pdf and run make.  It seems overkill to list them as say
GRAPHICS_FILES and make that list a dependency for sc21.pdf.  It seems
hard enough to remember to update the extras file, without the
additional burden of updating Makefile.am too.