SC/22: Put Stokes-parameter symbols consistently in italics.
Earlier parts of the document had Q, U, and I in italics as these are
symbols, but later additions dod not.  This follows the style of the
main astronomical journals.  Perhaps IP and ANG should also be
likewise modified, but I took IP as an abbreviation, and ANG as a
column name in the catalogue.