cupid: Use offset from initial peak as minimisation parameter
Prevously, the minimisation process used absolute peak position within
the whole array to specify the centre of the fitted model. Now, it uses
the offset of the model peak from the initial guess. This improves the
conditioning of the minimisation parameters, since the scale of the
offset will typically be a few pixels, whereas the scale of the whole
array could be anywhere from a few tens upto tens of thousands.