KAPPA: Improve appearance of picture showing palette & colour table.
The diagram was concertinaed, implying that the \unitlength changed at
some point in recent years, possibly when the document style changed.
To restore the appearance as was, this commit sets the unit length to
3pt for the picture environment.

As the braces do not span the appropriate widths in the HTML diagram,
a graphics file in a figure environment is probably needed to overcome
a latex2html deficiency here.