ndf: Remove standlone NDF lib from list of libs returned by adam linker script
Previous versions of the NDF adam linker script included both libndf_adam
and libndf in the returned list of libraries. But all functions required
to implement the whole library are contained within libndf_adam and so
libndf is not needed. Including libndf seems to cause problems on some
versions of macos because both libraries include a function with the
"constructor" attribute (ndf1GetCmdLine). To avoid this issue, remove
libndf from the list of libraries returned by the adam linker script.