smurf: makemap/skyloop/pol2map: Prevent IP correction being applied twice
On the first iteration, skyloop tells makemap to export the cleaned data.
It then uses the cleaned data as input on all subsequent iterations (and
tells makemap NOT to re-clean the data). This saves the time spent
cleaning the data on every iteration.

However, the makemap code that exports cleaned data was not placed
immediately after the cleaning of the data. In fact, IP correction
occurred after the data was cleaned but before it was exported. So the
exported cleaned data includes IP correction.

On subsequent iterations, the "doclean" makemap config parameter is set
to zero to tell makemap not to clean the input data (because it has
already been cleaned on the first iteration). However, the doclean
parameter only inhibits cleaning, and does NOT inhibit the application of
IP correction. This means that IP correction will be re-applied to the
data on all subsequent iterations, even though the cleaned input data
already includes IP correction. This means that the data that goes into
the map on all subsequent iterations has had much stroinger IP correction
(two times stronger) than it should have had.

This change fixes this nasty bug by moving the makemap code that exports
the cleaned data so that it is placed immediately after the data is
cleaned (i.e. before IP correction or any other change is made to the
data).

This means that all previous POL2 vector maps created using skyloop
(which is the standard method used by BISTRO) could potentially be wrong
- depending on how significant the IP correction is compared to the
polarisation of the source.

Bad news.......

This bug also breaks the recently added INITIALSKY parameter in skyloop,
but I've not heard of anyone using INITIALSKY so that is not so much of a
problem (I added it to aid my own investigations of the 450/850 um offset
in Q and U reported by Ray).