smurf: Correct bug in SCUBA-2 gap-filling algorithm
The old version attempted to do a fit to the data in a box at the start
or end of a gap, even if there were only two good values in the box. This
could lead to a spurious gradient and offset and consequent bad
interpolation of the gap. Now a fit is only done if there are at least
"minboix" good values in the box being fitted. If no fit is done, the
mean value in the box is used directly, instead of doing a fit and using
the end point of the fitted line.