cupid: Change gaussclumps weighting scheme
The Gaussclumps weighting scheme assigns a weight to each pixel to be
used when doing the least squares fit that determines the parameters of
each best fitting gaussian. It is supposed to down-weight pixels that
have a large contribution from neighbouring clumps. The old scheme
updated the weights as the fitting proceded. But this caused the fitter
to become confused, resulting in occasional very poor fits. The new
scheme calculates the weights once before the iterative fitted commences,
and then leaves the weights unchanged.