cupid: Re-write GaussClumps signal handling for multi-threaded context
The old Gaussclumps interupt handling doesn't work now that the fitting
process has been multi-threaded. It causes seg faults and unpredicictable
behaviour. So change it to be like the makemap interupt handling - the
signal handler now just sets a global variable and returns to where it
was called to continue what it was doing before the interupt. Later cupid
code then checks this global variable and exits cleanly, saving all
clumps found so far, if it is set. Previously, the interupt handler just
jumped immediately to the clean up code in cupidGaussClumps leaving the
thread code in an unpredictable state.