smurf: New makemap output parameter NOMFCF holds nominal FCF
The nominal beam FCF is calculated as the weighted mean of the nominal
FCFs associated with each contiguous chunk/observation. The weight for a
chunk is the sum of its exposure time contributions to the output
exposure time map. Thus if the input observations span a date at which
the nominal FCF changes, the NOMFCF value will be a weighted mean of the
two nominal FCF values.