smurf: pol2map: Allow for datasets that span a change in FCF
If parameter JY is TRUE, the I, Q, U, PI values in the catalogue are in
units of mJy/beam. The default FCF used to convert from pW to mJy/beam is
now determined from makemap, which uses a weighted average of the FCFs
for different dates. The weights represent the amount of data used for
each date.