smurf: pol2map.py: Prevent spurious history record being added to ipref map
Previously, if an existing NDF was supplied for parameter IPREF on the
command line but an error occurred before the IPREF parameter was
accessed, the ParSys cleanup function would add a history record to the
existing ipref NDF suggesting it had been created by pol2map. This commit
fixes this bug by changing the definition of the IPREF parameter so that
it initially expects an existing NDF. Then temporarily change its
"exists" property just before the parameter value is accessed to cater
for cases where the IPREF map is also the output total intensity map and
should therefore have a history record added.