starconf: correct datatype for F77 trailing string length arguments
gfortran V8 changed the data type for trailing character length arguments
from int to size_t. CNF assumed int for all versions of gfortran. On
linux this was harmless because extra space is introduced by gcc between
arguments to ensure 16 byte stack alignement and this padding seems to be
initialised to zero. Clang on MacOS also uses 16 byte stack alignement
but does not seem to initialise the extra padding to zero. This means
that the four bytes of padding which are used, together with the actual
4-byte int value, within the called fortran code to make up the 8-byte
size_t, can have random contents, resulting in random values being used
as the size_t that gives the character string length. These random
character lengths can cause spurious segfaults on MacOS.

This change causes size_t to be used as the data type for gfortran
version of 8 and above.