smurf: pol2map: Fix bug when coadding existing maps
A bug was introduced by the recently added facility for using a weighted
FCF when the input data spans an FCF change date. If the input data
consists of a set of previously created maps (rather than time stream
data ), the script would fail to get a nominal FCF and would then crash.

This commit restructures the determination and storage of the nominal FCF
to avoid this.