ndf: Use starperl in make_h
make_h previously used a script pygrep which used an unversioned
"python" command, which might not be available on all systems.
However it was introduced to replace "grep -P", the Perl regular
expression mode of grep, and we know that starperl will be
available, so we might as well just use that instead.