gaia: Update test for macOS in gaia.sh.in
We can't use the presence of DYLD_LIBRARY_PATH to detect macOS
in recent versions as this variable seems to be removed when
starting scripts - so try using uname instead.

Also 'test "$DYLD_LIBRARY_PATH" = ""' doesn't seem to work if the
variable is undefined.  Replace this with '[ -z "$DYLD_LIBRARY_PATH" ]'.