figaro: Exit fitgauss with bad status on failure
ORAC-DR seems to expect fitgauss to exit with bad status in order
to know that fitting failed so it should not read pointing results
from the output parameters.  However this did not seem to be happening.
Perhaps the bad status from SPD_WFGA could be allowed to come through,
but for now just check FITTED at the end and set status to SAI__ERROR
if necessary.