kaplibs: Do not zero-index kpgBmdnSec el and m0
"el" is described as the "total number of elements" and "m0"
is the initial value for "m" which is described as a "counter of
output values".  Therefore it seems inappropriate to decrement
them to "convert these two indices from one-based to zero-based".
They appear never to be used as indices.  This avoids the need for
special handling of the value -1 at the start of the while loop.