kaplibs: Match kpgBmdnSec looping in bad = 0 case
kpgBmdnSec has two copies of its main processing loop, for bad != 0
and bad = 0 cases.  This updates the looping control code for the
bad = 0 case to match that of the first case.  For example, this
ensures that loops begin with m = m0 and not at 0 so that each
thread does not process more elements than it should.  The altered
limits for multi-threading and double-check that cdim does not
exceed ndim are also included.