cfitsio: Update to v4.4.0
The naming of the shared library seems to have changed, e.g.
from: libcfitsio.10.@PACKAGE_VERSION@@SHLIB_SUFFIX@
to:   libcfitsio@SHLIB_SUFFIX@.10.@PACKAGE_VERSION@

Use the "make utils" command specified in the new README.md to build
the utilities.  Install fitsverify since it is now included here.

Remove the copy of the old License.txt file and install the new license
directly from the submodule.  Add a link to the source code as
it requests.