buildsupport: Add script to create index for docs
This script is intended to create an index.html file for the
docs directory based on the information given in the docs_list file.
It only includes documents for which an HTML or PDF version
is found.  It expects $STARLINK_DIR to be set, indicating the
location of the software on which to operate.