gaia: Only draw part of MOC overlapping the plot
Previously attempting to draw a large MOC (e.g. with many cells
outside the area of the displayed image) was problematic because it
would involve AST computing the boundary of the entire MOC.  In order
to avoid this, the "createPlot" function was updated so that it
(optionally) returns an AST region representing the area plotted.
Then before the MOC is drawn, the plot region is added to it in
"and" mode to obtain their intersection.

Note that the intersection will be performed at current resolution
of the MOC.  (Its MinOrder attribute is set equal to its MaxOrder to
ensure the computation is complete at this order.)  This might give
acceptable results provided that the MOC's resolution is sufficient
to reasonably represent the plot area.  If this is not the case,
it might be necessary to work on a MOC at higher order and/or use a
modified region with some padding around the plotted region.  Another
possible improvement could be to attempt to determine whether the
MOC is large enough to require restriction to the plot region or not.