KAPPA: Second attempt to make full height/width plotting region.
Setting the RJUST values to negative was an attempt to create
full-height (or width) plotting areas without altering the API.  While
the original attempt (Commit 0005859f47a26) worked to some extent, in
sub-one aspect ratio (portrait) plots, the shrinkage undermined this.

The feature was added to enable KAPPA:DISPLAY with the KEY option to
have the LUT displayed to the full height of the DATA picture, and
hence not be shrunk vertically; or also laterally, leading to a narrow
LUT ramp and sometimes slightly overprinted numerical annotation.

The main change is to stop the shrinkage after the first iteration,
which allows for the annotations.

While testing on different aspect ratios give the desired result,
there's a nagging feeling that it's not implemented correctly in
general.  That's not testable at present as this routine is only
called from KPG1_LUTKY.