gaia: Allow changing color/width of pos. of int.
The "position of interest" is used to indicate coordinates received
via the SAMP "coord.pointAt.sky" message.  This commit adds two entries
to the "Interop" menu allowing a choice of the color and width
of the circle used to show this position.

The "position_of_interest" method was updated to extract methods
"del_pos_interest_" and "plot_pos_interest_" so that a new method
"redraw_pos_interest_" could be added to update the marker when
the options are changed.  In order to allow this to be done,
the position (RA and Dec.) and units of the coordinates are
stored in separate variables, replacing the single variable
(holding "$ra,$dec") used to determine when the coordinates
have changed.