pcs: Remove use of TIOCSTI
This is not available by default since Linux 6.2 and was already
disabled where not available (e.g. Cygwin).  It appears that the
surrounding code (disabling echo, reading input, converting \n to \r)
is probably not useful without the simulated newline being sent to
the terminal input, so remove it too.