SC/20: Clarify determination of map rotation angle
The situation described (map rotation by MAP_PA) is a somewhat
specific example for when the map has for some reason been forced
upright in the same coordinate system as the observation.  Maps made
at odd angles should be a less common occurrence given the various
bug fixes related to crota parameters in the 2023A release.