smurf: Delay add/n of mapping to FS in FTS-2 case
In sc2ast_createwcs, when including FTS-2 port mappings, delay
the addition of the mapping to the frameset until it can be
processed by sc2ast_fts_image_port.  This allows the "ftsportmap"
Ident to be cleared from the mapping as retained by the frameset.
The previous approach of extracting the mapping from the frameset
using astGetMapping seemed to return a copy so that removing the
Ident did not remove it from the frameset.