kappa: Add clinplot parameter DOWNSAMPLE
Added a new parameter to clinplot to control whether or not the spectra
are downsampled prior to plotting.  This is to allow clinplot to be
used for undulating data (such as interferograms) where downsampling
can obscure the nature of the data.  Also corrected the logic for
downsampling to use 4 pixels per point in histogram modes to work as
described in the preceding comment (histogram modes are 1 and 6, not 2,
which is line).  The downsampling for histograms being applied in the
default (line) mode made it especially unhelpful for interferograms.

The default for the new parameter is TRUE, retaining the previous
behavior except for the correction to the logic regarding histogram
modes.