KAPPA: Add alignment options to SQORST via its CENTRE parameter.
This permits tiled surveys (e.g. CHIMPS, CLOGS) to use the same channel
centres in each tile.  Alignment can be to channel centre or edge.  It
uses the new subroutine in Commit 8704add9ce006.

Increment the KAPPA version so that ORAC-DR can detect whether this
feature is available in the running Starlink.